Refactor RTC server and ROS2 integration; improve image handling and add main execution script
